machine-learning/speech-to-text-realtime/speech-to-text-realtime-ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
torch==2.1.0
openai-whisper==20231117
sounddevice==0.4.6
python-multipart==0.0.6
websockets==12.0 